fix alarm handling
Time::HiRes was not used at all, this is now fixed.

To test for a working Time::HiRes::alarm it seems that a test import
of Time::HiRes::ualarm has to be done.

Time::HiRes is not required at all if the timeout looks like an
integer.

Fix alarm(@_) call --- it seems that alarm() evaluates the arguments
in scalar context, so previosly it ended always with alarm(1).